Create first version of LevelController
